requirements.txt
attrs~=22.2.0
Django~=4.1.6
asgiref~=3.6.0
sqlparse~=0.4.3
packaging~=23.0
python-dotenv~=0.21.1
pytest~=7.2.1
Faker~=17.0.0
djangorestframework~=3.14.0